+++ src/cuda/tensorFactories.cu
@@ -5,26 +5,26 @@
 
 namespace microtorch {
 
-__global__ void fill_kernel(size_t n, float *self, float value) {
-  size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
-  size_t stride = blockDim.x * gridDim.x;
-  for (size_t i = idx; i < n; i += stride) {
+__global__ void fill_kernel(int64_t n, float *self, float value) {
+  int64_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+  int64_t stride = blockDim.x * gridDim.x;
+  for (int64_t i = idx; i < n; i += stride) {
     self[i] = value;
   }
 }
 
 template <>
 void fill_impl<Cuda>(Tensor &self, const data_t value) {
-  size_t blocks_per_grid = get_blocks_per_grid(self.numel());
+  int64_t blocks_per_grid = get_blocks_per_grid(self.numel());
   fill_kernel<<<blocks_per_grid, ThreadsPerBlock>>>(self.numel(),
                                                     self.data_ptr(), value);
   CUDA_ERROR_CHECK();
 }
 
-__global__ void clone_kernel(size_t n, const float *a, float *out) {
-  size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
-  size_t stride = blockDim.x * gridDim.x;
-  for (size_t i = idx; i < n; i += stride) {
+__global__ void clone_kernel(int64_t n, const float *a, float *out) {
+  int64_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+  int64_t stride = blockDim.x * gridDim.x;
+  for (int64_t i = idx; i < n; i += stride) {
     out[i] = a[i];
   }
 }
@@ -34,16 +34,16 @@ void clone_impl<Cuda>(const Tensor &a, Tensor &out) {
   float *a_ptr = a.data_ptr();
   float *out_ptr = out.data_ptr();
 
-  size_t blocks_per_grid = get_blocks_per_grid(a.numel());
+  int64_t blocks_per_grid = get_blocks_per_grid(a.numel());
   clone_kernel<<<blocks_per_grid, ThreadsPerBlock>>>(a.numel(), a_ptr, out_ptr);
   CUDA_ERROR_CHECK();
 }
 
-__global__ void clone_backward_kernel(size_t n, const float *grad_output_ptr,
+__global__ void clone_backward_kernel(int64_t n, const float *grad_output_ptr,
                                       float *grad_input_ptr) {
-  size_t idx = blockIdx.x * blockDim.x + threadIdx.x;
-  size_t stride = blockDim.x * gridDim.x;
-  for (size_t i = idx; i < n; i += stride) {
+  int64_t idx = blockIdx.x * blockDim.x + threadIdx.x;
+  int64_t stride = blockDim.x * gridDim.x;
+  for (int64_t i = idx; i < n; i += stride) {
     grad_input_ptr[i] = grad_output_ptr[i];
   }
 }
@@ -53,7 +53,7 @@ void clone_backward_impl<Cuda>(const Tensor &grad_output, Tensor &grad_input) {
   float *grad_output_ptr = grad_output.data_ptr();
   float *grad_input_ptr = grad_input.data_ptr();
 
-  size_t blocks_per_grid = get_blocks_per_grid(grad_output.numel());
+  int64_t blocks_per_grid = get_blocks_per_grid(grad_output.numel());
   clone_backward_kernel<<<blocks_per_grid, ThreadsPerBlock>>>(
       grad_output.numel(), grad_output_ptr, grad_input_ptr);
   CUDA_ERROR_CHECK();
@@ -61,7 +61,7 @@ void clone_backward_impl<Cuda>(const Tensor &grad_output, Tensor &grad_input) {
 
 __global__ void rand_kernel(float *data, int numel) {
   int idx = threadIdx.x + blockIdx.x * blockDim.x;
-  size_t stride = blockDim.x * gridDim.x;
+  int64_t stride = blockDim.x * gridDim.x;
 
   // For each block, initialize it's own state
   __shared__ curandState state;
@@ -77,7 +77,7 @@ __global__ void rand_kernel(float *data, int numel) {
 
 template <>
 void rand_impl<Cuda>(Tensor &self) {
-  size_t blocks_per_grid = get_blocks_per_grid(self.numel());
+  int64_t blocks_per_grid = get_blocks_per_grid(self.numel());
   rand_kernel<<<blocks_per_grid, ThreadsPerBlock>>>(self.data_ptr(),
                                                     self.numel());
   CUDA_ERROR_CHECK();
